+++ jupiter-rpc/src/main/java/org/jupiter/rpc/consumer/GenericProxyFactory.java
@@ -206,7 +206,7 @@ public class GenericProxyFactory {
         }
 
         // dispatcher
-        Dispatcher dispatcher = dispatcher(metadata, serializerType)
+        Dispatcher dispatcher = dispatcher()
                 .hooks(hooks)
                 .timeoutMillis(timeoutMillis)
                 .methodSpecialConfigs(methodSpecialConfigs);
@@ -214,21 +214,21 @@ public class GenericProxyFactory {
         ClusterStrategyConfig strategyConfig = ClusterStrategyConfig.of(strategy, retries);
         switch (invokeType) {
             case SYNC:
-                return new SyncGenericInvoker(client, dispatcher, strategyConfig, methodSpecialConfigs);
+                return new SyncGenericInvoker(client.appName(), metadata, dispatcher, strategyConfig, methodSpecialConfigs);
             case ASYNC:
-                return new AsyncGenericInvoker(client, dispatcher, strategyConfig, methodSpecialConfigs);
+                return new AsyncGenericInvoker(client.appName(), metadata, dispatcher, strategyConfig, methodSpecialConfigs);
             default:
                 throw reject("invokeType: " + invokeType);
         }
     }
 
-    protected Dispatcher dispatcher(ServiceMetadata metadata, SerializerType serializerType) {
+    protected Dispatcher dispatcher() {
         switch (dispatchType) {
             case ROUND:
                 return new DefaultRoundDispatcher(
-                        LoadBalancerFactory.loadBalancer(loadBalancerType), metadata, serializerType);
+                        client, LoadBalancerFactory.loadBalancer(loadBalancerType), serializerType);
             case BROADCAST:
-                return new DefaultBroadcastDispatcher(metadata, serializerType);
+                return new DefaultBroadcastDispatcher(client, serializerType);
             default:
                 throw reject("dispatchType: " + dispatchType);
         }