+++ jupiter-rpc/src/main/java/org/jupiter/rpc/consumer/dispatcher/DefaultBroadcastDispatcher.java
@@ -24,7 +24,6 @@ import org.jupiter.rpc.consumer.future.DefaultInvokeFuture;
 import org.jupiter.rpc.consumer.future.DefaultInvokeFutureGroup;
 import org.jupiter.rpc.consumer.future.InvokeFuture;
 import org.jupiter.rpc.model.metadata.MessageWrapper;
-import org.jupiter.rpc.model.metadata.ServiceMetadata;
 import org.jupiter.serialization.Serializer;
 import org.jupiter.serialization.SerializerType;
 import org.jupiter.transport.channel.JChannel;
@@ -40,27 +39,18 @@ import org.jupiter.transport.channel.JChannelGroup;
  */
 public class DefaultBroadcastDispatcher extends AbstractDispatcher {
 
-    public DefaultBroadcastDispatcher(ServiceMetadata metadata, SerializerType serializerType) {
-        super(metadata, serializerType);
+    public DefaultBroadcastDispatcher(JClient client, SerializerType serializerType) {
+        super(client, serializerType);
     }
 
     @SuppressWarnings("unchecked")
     @Override
-    public <T> InvokeFuture<T> dispatch(JClient client, String methodName, Object[] args, Class<T> returnType) {
+    public <T> InvokeFuture<T> dispatch(JRequest request, Class<T> returnType) {
         // stack copy
-        final ServiceMetadata _metadata = metadata();
         final Serializer _serializer = serializer();
+        final MessageWrapper message = request.message();
 
-        MessageWrapper message = new MessageWrapper(_metadata);
-        message.setAppName(client.appName());
-        message.setMethodName(methodName);
-        // 不需要方法参数类型, 服务端会根据args具体类型按照JLS规则动态dispatch
-        message.setArgs(args);
-
-        JChannelGroup[] groups = client
-                .connector()
-                .directory(_metadata)
-                .snapshot();
+        JChannelGroup[] groups = groups(message.getMetadata());
         JChannel[] channels = new JChannel[groups.length];
         for (int i = 0; i < groups.length; i++) {
             channels[i] = groups[i].next();
@@ -69,15 +59,12 @@ public class DefaultBroadcastDispatcher extends AbstractDispatcher {
         byte s_code = _serializer.code();
         // 在业务线程中序列化, 减轻IO线程负担
         byte[] bytes = _serializer.writeObject(message);
-
-        JRequest request = new JRequest();
-        request.message(message);
         request.bytes(s_code, bytes);
 
         long invokeId = request.invokeId();
         ConsumerHook[] hooks = hooks();
         InvokeFuture<T>[] futures = new DefaultInvokeFuture[channels.length];
-        long timeoutMillis = getMethodSpecialTimeoutMillis(methodName);
+        long timeoutMillis = getMethodSpecialTimeoutMillis(message.getMethodName());
         for (int i = 0; i < channels.length; i++) {
             JChannel ch = channels[i];
             DefaultInvokeFuture<T> future = DefaultInvokeFuture