+++ jupiter-rpc/src/main/java/org/jupiter/rpc/consumer/dispatcher/Dispatcher.java
@@ -17,10 +17,9 @@
 package org.jupiter.rpc.consumer.dispatcher;
 
 import org.jupiter.rpc.ConsumerHook;
-import org.jupiter.rpc.JClient;
+import org.jupiter.rpc.JRequest;
 import org.jupiter.rpc.consumer.future.InvokeFuture;
 import org.jupiter.rpc.model.metadata.MethodSpecialConfig;
-import org.jupiter.rpc.model.metadata.ServiceMetadata;
 
 import java.util.List;
 
@@ -32,9 +31,7 @@ import java.util.List;
  */
 public interface Dispatcher {
 
-    <T> InvokeFuture<T> dispatch(JClient client, String methodName, Object[] args, Class<T> returnType);
-
-    ServiceMetadata metadata();
+    <T> InvokeFuture<T> dispatch(JRequest request, Class<T> returnType);
 
     Dispatcher hooks(List<ConsumerHook> hooks);