+++ jupiter-rpc/src/main/java/org/jupiter/rpc/consumer/invoker/AsyncGenericInvoker.java
@@ -16,13 +16,12 @@
 
 package org.jupiter.rpc.consumer.invoker;
 
-import org.jupiter.rpc.JClient;
-import org.jupiter.rpc.consumer.cluster.ClusterInvoker;
 import org.jupiter.rpc.consumer.dispatcher.Dispatcher;
 import org.jupiter.rpc.consumer.future.InvokeFuture;
 import org.jupiter.rpc.consumer.future.InvokeFutureContext;
 import org.jupiter.rpc.model.metadata.ClusterStrategyConfig;
 import org.jupiter.rpc.model.metadata.MethodSpecialConfig;
+import org.jupiter.rpc.model.metadata.ServiceMetadata;
 
 import java.util.List;
 
@@ -36,21 +35,22 @@ import java.util.List;
  *
  * @author jiachun.fjc
  */
-public class AsyncGenericInvoker extends ClusterStrategyBridging implements GenericInvoker {
+public class AsyncGenericInvoker extends AbstractInvoker implements GenericInvoker {
 
-    public AsyncGenericInvoker(JClient client,
+    public AsyncGenericInvoker(String appName,
+                               ServiceMetadata metadata,
                                Dispatcher dispatcher,
                                ClusterStrategyConfig defaultStrategy,
                                List<MethodSpecialConfig> methodSpecialConfigs) {
-
-        super(client, dispatcher, defaultStrategy, methodSpecialConfigs);
+        super(appName, metadata, dispatcher, defaultStrategy, methodSpecialConfigs);
     }
 
     @Override
     public Object $invoke(String methodName, Object... args) throws Throwable {
-        ClusterInvoker invoker = getClusterInvoker(methodName);
-        InvokeFuture<?> future = invoker.invoke(methodName, args, Object.class);
-        InvokeFutureContext.set(future);
+        Object result = doInvoke(methodName, args, Object.class, false);
+
+        InvokeFutureContext.set((InvokeFuture<?>) result);
+
         return null;
     }
 }